Worked out error message, added grid
