+++ client/src/components/YourOutfit/AddOutfitButton.jsx
@@ -24,28 +24,15 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-
+//add to card:
+//onClick={() => props.handleAddOutfitClick()}
 const AddOutfitButton = (props) => {
   const classes = useStyles();
 
-  const handleAddOutfitClick = () => {
-    const newOutfitList = props.outfitList.slice();
-    if (newOutfitList.length === 0) {
-      newOutfitList.push(props.currentProduct);
-      props.setOutfitList(newOutfitList);
-    }
-    for (var i = 0; i < newOutfitList.length; i++) {
-      var current = newOutfitList[i];
-      if (current.id !== props.currentProduct.id) {
-        newOutfitList.push(props.currentProduct);
-      }
-    }
-    props.setOutfitList(newOutfitList);
-  }
 
   return (
     <React.Fragment>
-      <Card className={classes.root} variant="outlined" onClick={() => handleAddOutfitClick()}>
+      <Card className={classes.root} variant="outlined" onClick={() => props.handleAddOutfitClick()}>
         <CardContent>
           <Typography className={classes.title} gutterBottom>
             Add to Your Outfit
